chore(ui-react-storage): align package architecture with new package structure
