--- packages/react-storage/src/StorageManager/Container/index.ts
@@ -1,1 +1,0 @@
-export { Container } from './Container';
