--- packages/react-storage/src/StorageManager/DropZone/index.ts
@@ -1,1 +1,0 @@
-export { DropZone } from './DropZone';
