--- packages/react-storage/src/StorageManager/FileList/index.ts
@@ -1,1 +1,0 @@
-export { FileList } from './FileList';
