--- packages/react-storage/src/StorageManager/FileListContainer/FileListContainer.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { View } from '@aws-amplify/ui-react';
-
-export interface ContainerProps {
-  children: React.ReactNode;
-  className: string;
-}
-
-export function FileListContainer({
-  children,
-  className,
-}: ContainerProps): JSX.Element {
-  return <View className={className}>{children}</View>;
-}
