--- packages/react-storage/src/StorageManager/FileListContainer/index.ts
@@ -1,1 +1,0 @@
-export { FileListContainer } from './FileListContainer';
