--- packages/react-storage/src/StorageManager/FileListHeader/index.ts
@@ -1,1 +1,0 @@
-export { FileListHeader } from './FileListHeader';
