--- packages/react-storage/src/StorageManager/StorageManager/index.ts
@@ -1,1 +1,0 @@
-export { StorageManager } from './StorageManager';
