+++ packages/react-storage/src/components/StorageManager/index.ts
@@ -1,1 +1,2 @@
 export { StorageManager } from './StorageManager';
+export { StorageManagerProps } from './types';
