+++ packages/react-storage/src/components/StorageManager/ui/Container/index.ts
@@ -1,0 +1,1 @@
+export { Container, ContainerProps } from './Container';
