+++ packages/react-storage/src/components/StorageManager/ui/DropZone/DropZone.tsx
