+++ packages/react-storage/src/components/StorageManager/ui/DropZone/index.ts
@@ -1,0 +1,3 @@
+export { DropZone } from './DropZone';
+export { FilePicker } from './FilePicker';
+export { DropZoneProps, FilePickerProps } from './types';
