+++ packages/react-storage/src/components/StorageManager/ui/FileList/FileDetails.tsx
