+++ packages/react-storage/src/components/StorageManager/ui/FileList/FileRemoveButton.tsx
