+++ packages/react-storage/src/components/StorageManager/ui/FileList/FileThumbnail.tsx
