+++ packages/react-storage/src/components/StorageManager/ui/FileList/index.ts
@@ -1,0 +1,2 @@
+export { FileList } from './FileList';
+export { FileListProps } from './types';
