+++ packages/react-storage/src/components/StorageManager/ui/FileListHeader/index.ts
@@ -1,0 +1,1 @@
+export { FileListHeader, FileListHeaderProps } from './FileListHeader';
