+++ packages/react-storage/src/components/StorageManager/ui/index.ts
@@ -1,0 +1,9 @@
+export { Container, ContainerProps } from './Container';
+export {
+  DropZone,
+  DropZoneProps,
+  FilePicker,
+  FilePickerProps,
+} from './DropZone';
+export { FileList, FileListProps } from './FileList';
+export { FileListHeader, FileListHeaderProps } from './FileListHeader';
