+++ packages/react-storage/src/components/StorageManager/utils/__tests__/humanFileSize.test.ts
@@ -1,7 +1,7 @@
 import { humanFileSize } from '../humanFileSize';
 
 describe('humanFileSize', () => {
-  it('returns the correct human readable string for 100000 bytes metric false ', () => {
+  it('returns the correct human readable string for 100000 bytes metric false', () => {
     const message = humanFileSize(100000);
 
     expect(message).toEqual('97.7 KiB');
