+++ packages/react-storage/src/components/StorageManager/utils/checkMaxFileSize.ts
