+++ packages/react-storage/src/components/StorageManager/utils/displayText.ts
