+++ packages/react-storage/src/components/StorageManager/utils/index.ts
@@ -1,0 +1,8 @@
+export { checkMaxFileSize } from './checkMaxFileSize';
+export {
+  defaultStorageManagerDisplayText,
+  StorageManagerDisplayText,
+} from './displayText';
+export { filterAllowedFiles } from './filterAllowedFiles';
+export { humanFileSize } from './humanFileSize';
+export { uploadFile } from './uploadFile';
