+++ packages/react-storage/src/components/StorageManager/utils/uploadFile.ts
