+++ packages/react-storage/src/components/index.ts
@@ -1,0 +1,1 @@
+export { StorageManager, StorageManagerProps } from './StorageManager';
